authenticate and authorize using JWT token

+++ controllers/api/v1/posts_api.js
@@ -26,31 +26,21 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
             return res.json(200, {
                 message: "Post and associated comments deleted successfullly!",
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!",
+                
+
             })
-
-            // if(req.xhr){
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id,
-            //         },
-            //         message:"post deleted",
-            //     })
-            // }
-
-            // req.flash('success', 'Post and associated comments deleted!');
-
-            // return res.redirect('back');
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
 
     }catch(err){
         // req.flash('error', err);
